+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Load Questionnaire from DB
 - Search by WHO identifier
 - Sync progress loader and datetime
+- ANC details activity
 
 ### Fixed
 
@@ -36,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Reduce font size and spacing on patient list screen
 - removed covax prefix and postfix in general use-case
+- Changed patient registration extraction to use StructureMap-based extraction
 
 ## [0.0.3] - 2021-05-24
 ### Added